chatbot/mistralai/requirements.txt
requests==2.32.5
black==25.12.0
PyYAML==6.0.3
termcolor==3.1.0
emoji==2.15.0
transformers==4.57.3
torch==2.2.2
numpy==1.26.4
langchain==1.2.2
langchain-classic==1.0.1
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.6
langchain-mistralai==1.1.1
isort==7.0.0
jsonlines==4.0.0
tiktoken==0.12.0
rich==14.2.0